requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
peewee==3.15.4
python-telegram-bot-calendar==1.0.5
Pillow==9.2.0
requests==2.28.1
